+++ src/main/java/com/juanbenevento/wms/inventory/domain/event/StockMovedEvent.java
@@ -1,11 +1,12 @@
 package com.juanbenevento.wms.inventory.domain.event;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record StockMovedEvent(
         String lpn,
         String sku,
-        Double quantity,
+        BigDecimal quantity,
         String oldLocation,
         String newLocation,
         String username,
